requirements.txt
python-binance
TA-Lib
numpy
backtrader
flask
websocket-client